+++ lib/src/assetid_utils.dart
@@ -7,7 +7,8 @@ import 'package:path/path.dart';
 // convert to posix
 String normalizePath(String path) {
   // Ugly...
-  return posix.normalize(posix.joinAll(posix.split(posix.joinAll(windows.split(path)))));
+  return posix.normalize(
+      posix.joinAll(posix.split(posix.joinAll(windows.split(path)))));
 }
 
 /// generate the target assetId for a given path