+++ lib/src/html_printer.dart
@@ -169,7 +169,7 @@ class PrinterLines extends DelegatingList<PrinterLine> {
   final List<PrinterLine> _l;
 
   PrinterLines() : this.from(<PrinterLine>[]);
-  PrinterLines.from( List<PrinterLine> l)
+  PrinterLines.from(List<PrinterLine> l)
       : _l = l,
         super(l);
 }
@@ -178,7 +178,7 @@ class NodeLines extends DelegatingList<NodeLine> {
   final List<NodeLine> _l;
 
   NodeLines() : this.from(<NodeLine>[]);
-  NodeLines.from( List<NodeLine> l)
+  NodeLines.from(List<NodeLine> l)
       : _l = l,
         super(l);
 }