+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Initial revision
 * Travis test support
 * Tidy html
+* Lightweight css preprocessor